==> scripts/10-Monomorphic_check.Rmd <==
---
title: "10-Monomorphic_check"
author: "Erin D'Agnese"
date: "2023-06-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load libraries}
library(here)
library(tidyverse)
```

```{r import datasets}
MonoSNPsList <- read.csv(here("ListOfMonomorphicAlleles.csv"), header = TRUE)
ProbeSeq <- read.csv(here("Processing_GTseq_data", "GTseqOptimizationRd3", "ProbeSeq.csv"), header = FALSE)


```
